Move read functions from "utils" to "input_reader".
