+++ pylib-v1/ailets/models/gpt4o/messages_to_query.py
@@ -7,7 +7,8 @@ from ailets.cons.atyping import (
     ContentItemFunction,
     INodeRuntime,
 )
-from ailets.cons.util import iter_input_objects, read_all, write_all
+from ailets.cons.input_reader import iter_input_objects, read_all
+from ailets.cons.util import write_all
 from ailets.models.gpt4o.lib.typing import Gpt4oContentItem, Gpt4oMessage
 
 url = "https://api.openai.com/v1/chat/completions"
